pyproject.toml
[project]
name = "screener-bot"
version = "0.1.0"
description = "Telegram bot for on-demand screener portfolio checks"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "python-telegram-bot[job-queue]>=21.0",
    "pyyaml>=6.0",
    "screener",
]

[project.scripts]
screener-bot = "screener_bot.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[tool.uv.sources]
screener = { git = "https://github.com/KarneeshkarV/screener.git" }
